refactor: 적당히 모듈화 & 디렉토리 조정
