+++ src/app.module.ts
@@ -1,37 +1,13 @@
-import {
-  type MiddlewareConsumer,
-  Module,
-  type NestModule,
-} from "@nestjs/common";
-import { ConfigModule } from "@nestjs/config";
-import { TypeOrmModule } from "@nestjs/typeorm";
-import { AppController } from "./app.controller";
+import { Module } from "@nestjs/common";
+
 import { AppService } from "./app.service";
-import { LoggerMiddleware } from "./logger.middleware";
-import { RecruitModule } from "./recruit/recruit.module";
-import { S3Module } from "./s3/s3.module";
+import { RecruitModule } from "./applications/recruit/recruit.module";
+import { SystemModule } from "@/common/system/system.module";
+import { AppController, RecruitController } from "@/presentations/http";
 
 @Module({
-  imports: [
-    ConfigModule.forRoot({ envFilePath: ".env", isGlobal: true }),
-    TypeOrmModule.forRoot({
-      type: "postgres",
-      host: process.env.DB_HOST,
-      port: Number(process.env.DB_PORT),
-      username: process.env.DB_USERNAME,
-      password: process.env.DB_PASSWORD || "5334",
-      database: process.env.DB_DATABASE,
-      autoLoadEntities: true,
-      logging: true,
-    }),
-    RecruitModule,
-    S3Module,
-  ],
-  controllers: [AppController],
+  imports: [SystemModule, RecruitModule],
+  controllers: [AppController, RecruitController],
   providers: [AppService],
 })
-export class AppModule implements NestModule {
-  configure(consumer: MiddlewareConsumer) {
-    consumer.apply(LoggerMiddleware).forRoutes("*");
-  }
-}
+export class AppModule {}
