+++ src/common/support/s3/s3.module.ts
@@ -1,10 +1,8 @@
 import { Module } from "@nestjs/common";
-import { S3Controller } from "./s3.controller";
 import { S3Service } from "./s3.service";
 
 @Module({
   providers: [S3Service],
-  controllers: [S3Controller],
   exports: [S3Service],
 })
 export class S3Module {}
