+++ src/common/system/system.module.ts
@@ -1,0 +1,27 @@
+import { LoggerMiddleware } from "@/common/system/middlewares/logger.middleware";
+import { Module, NestModule, MiddlewareConsumer } from "@nestjs/common";
+import { ConfigModule } from "@nestjs/config";
+import { TypeOrmModule } from "@nestjs/typeorm";
+
+@Module({
+  imports: [
+    ConfigModule.forRoot({ envFilePath: ".env", isGlobal: true }),
+    TypeOrmModule.forRoot({
+      type: "postgres",
+      host: process.env.DB_HOST,
+      port: Number(process.env.DB_PORT),
+      username: process.env.DB_USERNAME,
+      password: process.env.DB_PASSWORD || "5334",
+      database: process.env.DB_DATABASE,
+      autoLoadEntities: true,
+      logging: true,
+    }),
+  ],
+  controllers: [],
+  providers: [],
+})
+export class SystemModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer.apply(LoggerMiddleware).forRoutes("*");
+  }
+}
