+++ src/main.ts
@@ -3,8 +3,8 @@ import { NestFactory } from "@nestjs/core";
 
 import cors from "cors";
 import { AppModule } from "./app.module";
-import { HttpExceptionFilter } from "./httpException.filter";
-import { InternalErrorFilter } from "./internal-error.filter";
+import { HttpExceptionFilter } from "./common/system/filters/httpException.filter";
+import { InternalErrorFilter } from "./common/system/filters/internal-error.filter";
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
   app.useGlobalFilters(new InternalErrorFilter(), new HttpExceptionFilter());
@@ -12,7 +12,7 @@ async function bootstrap() {
     cors({
       origin: ["http://localhost:5173", "https://devkor.club"],
       // TODO: 배포 후 CORS 설정 변경
-    }),
+    })
   );
   app.useGlobalPipes(
     new ValidationPipe({
@@ -21,15 +21,15 @@ async function bootstrap() {
         const messages = errors
           .map((error) => {
             return `<${error.property}> ${Object.values(error.constraints).join(
-              " ",
+              " "
             )}`;
           })
           .join(" ");
         return new NotAcceptableException(
-          `입력값이 유효하지 않습니다 - ${messages}`,
+          `입력값이 유효하지 않습니다 - ${messages}`
         );
       },
-    }),
+    })
   );
   await app.listen(3071);
 }
