+++ src/presentations/http/app/dto/presignedUrl.dto.ts
