+++ src/presentations/http/index.ts
@@ -1,0 +1,2 @@
+export * from "./app/app.controller";
+export * from "./recruit/recruit.controller";
