--- src/s3/dto/index.ts
@@ -1,1 +1,0 @@
-export * from "./presignedUrl.dto";
